// cypress/PageDetail/enums/enums.ts
export enum Timeout{
    SHORT = 2000,
    MEDIUM = 5000,
    LONG = 10000
}
export enum FlightLocation {
      FROM = 'from',
      TO = 'to'
}
export enum AirportName {
    DEPARTURE = "ln",
    DESTINATION = "lg",
}
export enum selectTripDates{
    From = 0,
    To = 1
} 
export enum CalculatedDays{
    // 60 132
    DepartureDays = 60,
    ArrivalDays = 132   
}